third_party/blink/web_tests/external/wpt/css/geometry/DOMMatrix-invertSelf-expected.txt

This is a testharness.js-based test.
[FAIL] new DOMMatrix([1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1]).invertSelf()
  assert_equals: Expected value for m12 is 0 expected 0 but got -0
Harness: the test ran to completion.
